prefect/collect_matches/requirements.txt
requests
pandas==1.5.2
prefect==2.9.0
prefect-gcp[cloud_storage]==0.2.4
protobuf==4.21.11
pyarrow==10.0.1
pandas-gbq==0.18.1
google-cloud-storage
fsspec
gcsfs
pydantic==1.10.11
